refactor(cmd): move openFileInEditor util to utils file
